+++ platform/apple/VpnClient/Logger.h
@@ -16,6 +16,10 @@ typedef void (^NativeLoggerCallback)(LogLevel log_level, NSString *message);
 
 + (void)setCallback:(nullable NativeLoggerCallback)callback;
 
++ (LogLevel)currentLogLevel;
+
++ (void)setLogLevel:(LogLevel)logLevel;
+
 @end
 
 NS_ASSUME_NONNULL_END