+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 
+- Respect log level from config for apple and android adapters.
+
 ### Security
 
 ## [1.1.5-beta.12] - 2026-07-08